[DOCS] 카카오 로그인시 예외 처리 클래스 추가 - OAuth 예외 처리 클래스 통합
