--- src/main/java/com/effourt/calenkit/exception/handler/OauthExceptionHandler.java
@@ -1,23 +1,0 @@
-package com.effourt.calenkit.exception.handler;
-
-import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
-import org.springframework.security.oauth2.core.OAuth2AuthorizationException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@ControllerAdvice
-public class OauthExceptionHandler {
-
-    @ExceptionHandler(OAuth2AuthenticationException.class)
-    public void authenticationEx(HttpServletResponse response) throws IOException {
-        response.sendError(HttpServletResponse.SC_FORBIDDEN);
-    }
-
-    @ExceptionHandler(OAuth2AuthorizationException.class)
-    public void authorizationEx(HttpServletResponse response) throws IOException {
-        response.sendError(HttpServletResponse.SC_UNAUTHORIZED);
-    }
-}
